Fix hdfs.init() when called on a Windows machine with the working directory a UNC path
On Windows (tested on Windows 7), if the current working directory is
a UNC path, say '\\host\share$\user\', the CMD.EXE interpreter returns
a warning message on running hadoop.cmd with argument 'classpath'. The
full output of the system call is [text translated from German]:

\\host\share$\user\
CMD.EXE was called with the above path as current directory.
UNC paths are not supported.
The Windows directory will be used as current directory instead.
C:\Progra~1\Hadoop\hadoop-2.7.3\etc\hadoop;[more paths...]

On the last line, the desired classpath is given, so hdfs.init()
should use the last line of the output of the system call to
hadoop.cmd instead of the first.